+++ search/query/disjunction.go
@@ -15,15 +15,10 @@
 package query
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 
-	"github.com/blevesearch/bleve/v2/mapping"
-	"github.com/blevesearch/bleve/v2/search"
-	"github.com/blevesearch/bleve/v2/search/searcher"
 	"github.com/blevesearch/bleve/v2/util"
-	index "github.com/blevesearch/bleve_index_api"
 )
 
 type DisjunctionQuery struct {
@@ -63,55 +58,6 @@ func (q *DisjunctionQuery) AddQuery(aq ...Query) {
 
 func (q *DisjunctionQuery) SetMin(m float64) {
 	q.Min = m
-}
-
-func (q *DisjunctionQuery) Searcher(ctx context.Context, i index.IndexReader, m mapping.IndexMapping,
-	options search.SearcherOptions) (search.Searcher, error) {
-	ss := make([]search.Searcher, 0, len(q.Disjuncts))
-	knnSearcherCount := 0
-	matchNoneSearcherCount := 0
-	for _, disjunct := range q.Disjuncts {
-		sr, err := disjunct.Searcher(ctx, i, m, options)
-		if err != nil {
-			for _, searcher := range ss {
-				if searcher != nil {
-					_ = searcher.Close()
-				}
-			}
-			return nil, err
-		}
-		if sr != nil {
-			if _, ok := sr.(*searcher.MatchNoneSearcher); ok && q.queryStringMode {
-				// in query string mode, skip match none
-				continue
-			}
-			if _, ok := sr.(*searcher.KNNSearcher); ok {
-				knnSearcherCount += 1
-			}
-			if _, ok := sr.(*searcher.MatchNoneSearcher); ok {
-				matchNoneSearcherCount += 1
-			}
-			ss = append(ss, sr)
-		}
-	}
-
-	if len(ss) < 1 {
-		return searcher.NewMatchNoneSearcher(i)
-	}
-
-	nctx := context.WithValue(ctx, search.IncludeScoreBreakdownKey, q.retrieveScoreBreakdown)
-
-	// Here, check if it's effectively a single KNN searcher
-	// If so, return new KNN Searcher
-	if knnSearcherCount == 1 && matchNoneSearcherCount == 1 {
-		finalSearchers, err := searcher.OptimizeKNNSearchers(nctx, i, ss, options)
-		if err != nil {
-			return nil, err
-		}
-		return finalSearchers[0], nil
-	}
-
-	return searcher.NewDisjunctionSearcher(nctx, i, ss, q.Min, options)
 }
 
 func (q *DisjunctionQuery) Validate() error {
